Added the ability to check if CouchDB is installed on a remote server. Added the ability to create a CouchDb admin user by default. Cleaned up some messages and powershell syntax

+++ id_auth/Install-Fabric.ps1
@@ -17,6 +17,30 @@ Import-Module -Name .\Fabric-Install-Utilities.psm1
 Add-Type -AssemblyName System.IO.Compression.FileSystem
 $workingDirectory = Split-Path $script:MyInvocation.MyCommand.Path
 
+function Get-CouchDbRemoteInstallationStatus($couchDbServer, $minVersion)
+{
+    try
+    {
+        $couchVersionResponse = Invoke-RestMethod -Method Get -Uri $couchDbServer 
+    } catch {
+        Write-Host "CouchDB not found on $couchDbServer"
+    }
+
+    if($couchVersionResponse)
+    {
+        $installedVersion = [System.Version]$couchVersionResponse.version
+        $minVersionAsSystemVersion = [System.Version]$minVersion
+        Write-Host "Found CouchDB version $installedVersion installed on $couchDbServer"
+        if($installedVersion -ge $minVersionAsSystemVersion)
+        {
+            return "Installed"
+        }else {
+            return "MinVersionNotMet"
+        }
+    }
+    return "NotInstalled"
+}
+
 function Get-FabricRelease($appName, $appVersion)
 {
     $outFile = "$env:Temp\Fabric.$appName.$appVersion.zip"
@@ -66,7 +90,7 @@ function Get-AccessToken($authUrl, $clientId, $scope, $secret)
     return $accessTokenResponse.access_token
 }
 
-function Create-ApiRegistration($authUrl, $body, $accessToken)
+function Add-ApiRegistration($authUrl, $body, $accessToken)
 {
     $url = "$authUrl/api/apiresource"
     $headers = @{"Accept" = "application/json"}
@@ -77,7 +101,7 @@ function Create-ApiRegistration($authUrl, $body, $accessToken)
     return $registrationResponse.apiSecret    
 }
 
-function Create-ClientRegistration($authUrl, $body, $accessToken)
+function Add-ClientRegistration($authUrl, $body, $accessToken)
 {
     $url = "$authUrl/api/client"
     $headers = @{"Accept" = "application/json"}
@@ -90,8 +114,9 @@ function Create-ClientRegistration($authUrl, $body, $accessToken)
 
 if(!(Test-Prerequisite '*.NET Core*Windows Server Hosting*' 1.1.30327.81))
 {
-    Write-Host "Windows Server Hosting Bundle minimum version 1.1.30327.81 not installed...installing version 1.1.30327.81"
+    Write-Host ".NET Core Windows Server Hosting Bundle minimum version 1.1.30327.81 not installed...downloading version 1.1.30327.81"
     Invoke-WebRequest -Uri https://go.microsoft.com/fwlink/?linkid=844461 -OutFile $env:Temp\bundle.exe
+    Write-Host "Installing .NET Core Windows Server Hosting Bundle..."
     Start-Process $env:Temp\bundle.exe -Wait -ArgumentList '/quiet /install'
     net stop was /y
     net start w3svc
@@ -101,18 +126,37 @@ if(!(Test-Prerequisite '*.NET Core*Windows Server Hosting*' 1.1.30327.81))
 if(!(Test-Prerequisite '*CouchDB*'))
 {
     #add check to see if couchdb url is reachable
-    Write-Host "CouchDB not installed, installing CouchDB Version 2.1.0.0"
-    Invoke-WebRequest -Uri https://dl.bintray.com/apache/couchdb/win/2.1.0/apache-couchdb-2.1.0.msi -OutFile $env:Temp\apache-couchdb-2.1.0.msi
-    Start-Process $env:Temp\apache-couchdb-2.1.0.msi -Wait
-    Remove-Item $env:Temp\apache-couchdb-2.1.0.msi
-    #add ability to create admin user
+    Write-Host "CouchDB not installed locally, testing to see if is installed on a remote server using $couchDbServer"
+    $remoteInstallationStatus = Get-CouchDbRemoteInstallationStatus $couchDbServer 2.0.0
+    if($remoteInstallationStatus -eq "NotInstalled")
+    {
+        Write-Host "CouchDB not installed, downloading CouchDB Version 2.1.0.0"
+        Invoke-WebRequest -Uri https://dl.bintray.com/apache/couchdb/win/2.1.0/apache-couchdb-2.1.0.msi -OutFile $env:Temp\apache-couchdb-2.1.0.msi
+        Write-Host "Launching CouchDB interactive installation..."
+        Start-Process $env:Temp\apache-couchdb-2.1.0.msi -Wait
+        Remove-Item $env:Temp\apache-couchdb-2.1.0.msi
+        $couchDbServer = "http://127.0.0.1:5984"
+        try{
+            Invoke-RestMethod -Method Put -Uri "$couchDbServer/_node/couchdb@localhost/_config/admins/$couchDbUsername" -Body "`"$couchDbPassword`""
+        } catch{
+            $exception = $_.Exception
+            Write-Host "Failed to create admin user for CouchDB. Exception $exception"
+        }
+
+    }elseif($remoteInstallationStatus -eq "MinVersionNotMet"){
+        Write-Host "CouchDB is installed on $couchDbServer but does not meet the minimum version requirements, you must have CouchDB 2.0.0.1 or greater installed: https://dl.bintray.com/apache/couchdb/win/2.1.0/apache-couchdb-2.1.0.msi"
+        exit 1
+    }else{
+        Write-Host "CouchDB installed and meets specifications"
+    }
 }elseif (!(Test-Prerequisite '*CouchDB*' 2.0.0.1)) {
     Write-Host "CouchDB is installed but does not meet the minimum version requirements, you must have CouchDB 2.0.0.1 or greater installed: https://dl.bintray.com/apache/couchdb/win/2.1.0/apache-couchdb-2.1.0.msi"
     exit 1
+}else{
+    Write-Host "CouchDB installed and meets specifications"
 }
 
 $identityServerUrl = "$hostUrl/identity"
-$authorizationServerUrl = "$hostUrl/authorization"
 
 $identityZipPackage = Get-FabricRelease Identity $identityVersion
 $identityArgs = Get-CommonArguments
@@ -120,7 +164,7 @@ $identityArgs += ("-zipPackage", $identityZipPackage)
 $identityArgs += ("-hostUrl", $hostUrl)
 Install-FabricRelease Identity $identityArgs
 
-cd $workingDirectory
+Set-Location $workingDirectory
 
 $authorizationZipPackage = Get-FabricRelease Authorization $authorizationVersion
 $authorizationArgs = Get-CommonArguments
@@ -128,7 +172,7 @@ $authorizationArgs += ("-zipPackage", $authorizationZipPackage)
 $authorizationArgs += ("-identityServerUrl", $identityServerUrl)
 Install-FabricRelease Authorization $authorizationArgs
 
-cd $workingDirectory
+Set-Location $workingDirectory
 
 #Register registration api
 $body = @'
@@ -138,7 +182,7 @@ $body = @'
     "scopes":[{"name":"fabric/identity.manageresources"}]
 }
 '@
-$registrationApiSecret = Create-ApiRegistration -authUrl $identityServerUrl -body $body
+$registrationApiSecret = Add-ApiRegistration -authUrl $identityServerUrl -body $body
 
 #Register Fabric.Installer
 $body = @'
@@ -150,7 +194,7 @@ $body = @'
     "allowedScopes": ["fabric/identity.manageresources", "fabric/authorization.read", "fabric/authorization.write", "fabric/authorization.manageclients"]
 }
 '@
-$installerClientSecret = Create-ClientRegistration -authUrl $identityServerUrl -body $body
+$installerClientSecret = Add-ClientRegistration -authUrl $identityServerUrl -body $body
 
 $accessToken = Get-AccessToken -authUrl $identityServerUrl -clientId "fabric-installer" -scope "fabric/identity.manageresources" -secret $installerClientSecret
 
@@ -162,7 +206,7 @@ $body = @'
     "scopes":[{"name":"fabric/authorization.read"}, {"name":"fabric/authorization.write"}, {"name":"fabric/authorization.manageclients"}]
 }
 '@
-$authorizationApiSecret = Create-ApiRegistration -authUrl $identityServerUrl -body $body -accessToken $accessToken
+$authorizationApiSecret = Add-ApiRegistration -authUrl $identityServerUrl -body $body -accessToken $accessToken
 
 #Register group fetcher client
 $body = @'
@@ -174,7 +218,7 @@ $body = @'
     "allowedScopes": ["fabric/authorization.read", "fabric/authorization.write", "fabric/authorization.manageclients"]
 }
 '@
-$groupFetcherSecret = Create-ClientRegistration -authUrl $identityServerUrl -body $body -accessToken $accessToken
+$groupFetcherSecret = Add-ClientRegistration -authUrl $identityServerUrl -body $body -accessToken $accessToken
 
 Write-Host "Please keep the following secrets in a secure place:"
 Write-Host "Fabric.Installer clientSecret: $installerClientSecret"
